requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.1
daphne==2.5.0
dataclasses==0.7
Django==3.0.8
django-cors-headers==3.4.0
django-countries==6.1.3
django-jazzmin==2.2.5
django-jsoneditor==0.1.6
django-language-field==0.0.3
django-multiselectfield==0.1.12
django-ninja==0.3.0
django-redis==4.12.1
djangorestframework==3.11.0
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
kombu==4.6.11
msgpack==1.0.0
packaging==20.4
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytz==2020.1
redis==3.5.3
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
stripe==2.51.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
vine==1.3.0
whitenoise==5.1.0
zipp==3.1.0
zope.interface==5.1.0
